requirements.txt
pyminder
pyyaml
